Move printers to a separate module
