+++ src/base/base-mixin.js
@@ -1,11 +1,12 @@
 import * as d3 from 'd3';
 
-import {pluck, printers, utils} from '../core/utils';
+import {pluck, utils} from '../core/utils';
 import {instanceOfChart, redrawAll, renderAll} from '../core/core';
 import {deregisterChart, registerChart} from '../core/chart-registry';
 import {constants} from '../core/constants';
 import {events} from '../core/events';
 import {logger} from '../core/logger';
+import {printers} from '../core/printers';
 import {InvalidStateException} from '../core/invalid-state-exception';
 import {BadArgumentException} from '../core/bad-argument-exception';
