+++ src/index.js
@@ -9,6 +9,7 @@ export * from './core/logger';
 export * from './core/config';
 export * from './core/events';
 export * from './core/filters';
+export * from './core/printers';
 export * from './core/units';
 export * from './base/d3.box';
